requirements.txt
asgiref==3.5.0
autopep8==1.6.0
boto3==1.23.7
botocore==1.26.7
Brotli==1.0.9
cffi==1.15.0
cssselect2==0.6.0
Django==4.0.3
django-ckeditor==6.4.0
django-environ==0.8.1
django-js-asset==2.0.0
django-storages==1.12.3
fonttools==4.33.3
html5lib==1.1
jmespath==1.0.0
numpy==1.23.1
pandas==1.4.3
Pillow==9.1.0
psycopg2-binary==2.9.3
pycodestyle==2.9.1
pycparser==2.21
pydyf==0.1.2
pyphen==0.12.0
python-dateutil==2.8.2
pytz==2022.1
s3transfer==0.5.2
six==1.16.0
sqlparse==0.4.2
tinycss2==1.1.1
toml==0.10.2
urllib3==1.26.9
weasyprint==54.3
webencodings==0.5.1
zopfli==0.2.1
